backend/requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
aiohttp==3.9.5
aiomysql==0.2.0
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
Django==5.0.7
django-cors-headers==4.4.0
django-filter==24.2
django-modeltranslation==0.19.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
environs==11.0.0
frozenlist==1.4.1
geographiclib==2.0
geopy==2.4.1
idna==3.7
inflection==0.5.1
magic-filter==1.0.12
marshmallow==3.21.3
multidict==6.0.5
mysqlclient==2.2.4
packaging==24.1
pillow==10.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
PyMySQL==1.1.1
pypng==0.20220715.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
whitenoise==6.7.0
yarl==1.9.4
